pyproject.toml
[project]
name = "fitnessapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.10",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.2.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
]
